Move code from 'Completions' to 'PathCompletions' namespace
